Styles restructuring: separated common styles, sprinkles; removed unused stuff from sprinkles and simplified responsive properties
