+++ src/components/sprinkles.css.ts
@@ -2,30 +2,24 @@ import { createTheme } from "@vanilla-extract/css";
 import { defineProperties, createSprinkles } from "@vanilla-extract/sprinkles";
 
 const padding = {
-  none: "0",
   extraSmall: "0.3rem",
   small: "0.6rem",
   medium: "0.8rem",
   large: "1rem",
-  extraLarge: "1.5rem",
-  // @deprecated
-  smallOld: "20px",
-  mediumOld: "40px",
+  extraLarge: "2rem",
 };
 
 const responsiveProperties = defineProperties({
   conditions: {
-    xs: {},
-    sm: { "@media": "screen and (min-width: 768px)" },
-    md: { "@media": "screen and (min-width: 1024px)" },
-    lg: { "@media": "screen and (min-width: 1400px)" },
-    xl: { "@media": "screen and (min-width: 2000px)" },
+    sm: { "@media": "screen and (max-width: 767px)" },
+    md: { "@media": "screen and (min-width: 768px)" },
+    lg: { "@media": "screen and (min-width: 1024px)" },
   },
   defaultCondition: "lg",
   properties: {
-    display: ["none", "flex", "block"],
     flexDirection: ["row", "column"],
-    justifyContent: ["stretch", "flex-start", "center", "flex-end", "space-around", "space-between"],
+    flexGrow: [1, 2],
+    flexBasis: ["10rem", "15rem", "20rem"],
     alignItems: ["stretch", "flex-start", "center", "flex-end"],
     paddingTop: padding,
     paddingBottom: padding,
@@ -34,31 +28,13 @@ const responsiveProperties = defineProperties({
     marginTop: [0, "0.5rem", "1rem", "2rem"],
     marginLeft: [0, "0.5rem", "1rem", "2rem"],
     height: ["auto", "30px", "50px", "70px", "100px", "150px", "200px"],
-    minHeight: ["auto", "30px", "50px", "70px", "100px", "150px", "200px"],
-    width: ["auto", "0.5rem", "1rem", "2rem",  "52px", "87px", "173px", "260px", "300px"],
-    minWidth: ["auto", "52px", "87px", "173px", "260px", "300px"],
+    width: ["1rem", "2rem", "87px", "173px", "260px"],
     fontSize: ["0.6rem", "0.8rem", "1rem", "1.2rem", "1.5rem", "2rem", "3rem"],
-    maxWidth: ["75rem", "100rem", "150rem"],
   },
   shorthands: {
     padding: ["paddingTop", "paddingBottom", "paddingLeft", "paddingRight"],
     paddingX: ["paddingLeft", "paddingRight"],
     paddingY: ["paddingTop", "paddingBottom"],
-    placeItems: ["justifyContent", "alignItems"],
-  },
-});
-
-export const [themeClass, vars] = createTheme({
-  color: {
-    white: "#ffffff",
-    red: "#ed1b24",
-    trueRed: "#cc0000",
-    black: "#333333",
-  },
-  shadows: {
-    small: "0 3px 6px 1px #9494944d",
-    medium: "0 3px 6px 3px #9494944d",
-    large: "0 3px 6px 5px #9494944d",
   },
 });
