+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import { AppProps } from "next/app";
 import { appWithTranslation } from "next-i18next";
-import { themeClass } from "../components/sprinkles.css";
+import { themeClass } from "../components/styles.css";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
